requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
faiss-cpu==1.10.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pip==25.0.1
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0